feat(revert): remove Primitive constraint on toThunk
The original purpose of Primitive constraint was to limit the scope of function usage,
reducing the likelihood of users passing multiple expressions as arguments to `toThunk`.
However, restricting the types that can be passed as arguments to only Primitive does
not prevent passing multiple expressions composed solely of Primitive values.
Therefore, the unclear type constraint associated with `toThunk` has been removed, and
the purpose of `toThunk` (evaluate the given expression FIRST and wrap it into Thunk)
has been clarified.